feat: refactor code out of store.go into other files
