+++ backend/db/sqlc/tx_multiple_stock_update.go
@@ -1,0 +1,1 @@
+package db
