+++ backend/db/sqlc/tx_sell_stock.go
@@ -1,0 +1,1 @@
+package db
